chore: dudeplicate code in cli
